Sửa nhận dien khuon mat

+++ ai-service/src/main/java/com/chamcongtinhluong/ai_service/dto/request/FaceApiRequest.java
@@ -13,5 +13,5 @@ import lombok.NoArgsConstructor;
 @Builder
 public class FaceApiRequest {
     private String employeeId;
-    private Integer[] faceDescriptor;
+    private Float[] faceDescriptor;
 }
